refactor: move JSONObjectHolder to libAnki module
* some extension methods moved to `PythonExtensions`

Removes a cyclic dependency: libAnki <-> AnkiDroid

Issue 18015
