+++ AnkiDroid/src/main/java/com/ichi2/libanki/CardTemplate.kt
@@ -16,8 +16,8 @@
 
 package com.ichi2.libanki
 
+import com.ichi2.anki.common.json.JSONObjectHolder
 import com.ichi2.anki.common.utils.ext.deepClone
-import com.ichi2.utils.JSONObjectHolder
 import net.ankiweb.rsdroid.RustCleanup
 import org.json.JSONObject
