+++ AnkiDroid/src/main/java/com/ichi2/libanki/Fields.kt
@@ -16,7 +16,7 @@
 
 package com.ichi2.libanki
 
-import com.ichi2.utils.JSONContainer
+import com.ichi2.anki.common.json.JSONContainer
 import org.json.JSONArray
 import org.json.JSONObject
