+++ AnkiDroid/src/test/java/com/ichi2/testutils/JsonUtils.kt
@@ -16,7 +16,7 @@
 
 package com.ichi2.testutils
 
-import com.ichi2.utils.JSONObjectHolder
+import com.ichi2.anki.common.json.JSONObjectHolder
 import org.hamcrest.BaseMatcher
 import org.hamcrest.Description
 import org.json.JSONObject
